- Create stacked_core package - Move annotation code to stacked_core
