--- packages/stacked/lib/stacked_annotations.dart
@@ -1,6 +1,0 @@
-export 'src/code_generation/stacked_router_annotations.dart';
-export 'src/code_generation/stacked_locator_annotations.dart';
-export 'src/code_generation/stacked_logger_annotations.dart';
-export 'src/code_generation/forms/stacked_form_annotations.dart';
-export 'src/code_generation/stacked_app.dart';
-export 'src/code_generation/environment_filter.dart';
