+++ packages/stacked_core/example/stacked_core_example.dart
@@ -1,0 +1,1 @@
+void main() {}
