+++ packages/stacked_core/lib/stacked_core.dart
@@ -1,0 +1,13 @@
+/// Support for doing something awesome.
+///
+/// More dartdocs go here.
+library stacked_core;
+
+export 'src/code_generation/stacked_router_annotations.dart';
+export 'src/code_generation/stacked_locator_annotations.dart';
+export 'src/code_generation/stacked_logger_annotations.dart';
+export 'src/code_generation/forms/stacked_form_annotations.dart';
+export 'src/code_generation/stacked_app.dart';
+export 'src/code_generation/environment_filter.dart';
+
+// TODO: Export any libraries intended for clients of this package.
