+++ packages/stacked_core/pubspec.yaml
@@ -1,0 +1,15 @@
+name: stacked_core
+description: Shared code between the stacked packages.
+version: 1.0.0
+# homepage: https://www.example.com
+
+environment:
+  sdk: '>=2.16.2 <3.0.0'
+
+
+# dependencies:
+#   path: ^1.8.0
+
+dev_dependencies:
+  lints: ^1.0.0
+  test: ^1.16.0
