+++ packages/stacked_core/test/stacked_core_test.dart
@@ -1,0 +1,5 @@
+import 'package:test/test.dart';
+
+void main() {
+  group('A group of tests', () {});
+}
