+++ packages/stacked_services/lib/src/navigation_service.dart
@@ -1,5 +1,4 @@
 import 'package:flutter/material.dart';
-import 'package:get/get.dart';
 import 'package:stacked_services/stacked_services.dart';
 
 class NavigationTransition {
