+++ packages/stacked_themes/test/test_setup.mocks.dart
@@ -1,22 +1,26 @@
-// Mocks generated by Mockito 5.0.9 from annotations
+// Mocks generated by Mockito 5.1.0 from annotations
 // in stacked_themes/test/test_setup.dart.
 // Do not manually edit this file.
 
 import 'dart:async' as _i5;
 import 'dart:ui' as _i6;
 
-import 'package:flutter/src/material/app.dart' as _i3;
+import 'package:flutter/material.dart' as _i3;
 import 'package:mockito/mockito.dart' as _i1;
 import 'package:stacked_themes/src/services/platform_service.dart' as _i7;
 import 'package:stacked_themes/src/services/shared_preferences_service.dart'
     as _i2;
 import 'package:stacked_themes/src/services/statusbar_service.dart' as _i4;
 
+// ignore_for_file: type=lint
 // ignore_for_file: avoid_redundant_argument_values
+// ignore_for_file: avoid_setters_without_getters
 // ignore_for_file: comment_references
+// ignore_for_file: implementation_imports
 // ignore_for_file: invalid_use_of_visible_for_testing_member
 // ignore_for_file: prefer_const_constructors
 // ignore_for_file: unnecessary_parenthesis
+// ignore_for_file: camel_case_types
 
 /// A class which mocks [SharedPreferencesService].
 ///
@@ -45,6 +49,12 @@ class MockStatusBarService extends _i1.Mock implements _i4.StatusBarService {
   _i5.Future<dynamic>? updateStatusBarColor(_i6.Color? statusBarColor) =>
       (super.noSuchMethod(
               Invocation.method(#updateStatusBarColor, [statusBarColor]))
+          as _i5.Future<dynamic>?);
+  @override
+  _i5.Future<dynamic>? updateNavigationBarColor(
+          _i6.Color? navigationBarColor) =>
+      (super.noSuchMethod(Invocation.method(
+              #updateNavigationBarColor, [navigationBarColor]))
           as _i5.Future<dynamic>?);
 }
